Move codebook to its own package.
This allows us to modularize it like experiment and imagestack.

Test plan: make -j fast run_notebooks
