+++ notebooks/py/DARTFISH_Pipeline_-_Human_Occipital_Cortex_-_1_FOV.py
@@ -21,8 +21,6 @@ import os
 from starfish import data
 from starfish.types import Features, Indices
 
-from starfish.codebook import Codebook
-
 from starfish.intensity_table import IntensityTable
 
 from starfish.image import Filter
