+++ starfish/util/synthesize.py
@@ -2,7 +2,7 @@ from typing import Tuple
 
 import numpy as np
 
-from starfish.codebook import Codebook
+from starfish.codebook.codebook import Codebook
 from starfish.imagestack.imagestack import ImageStack
 from starfish.intensity_table import IntensityTable
